update CI job to use the main branch
Update the CI jobs to use the main branch instead of the master
branch.  This should be merged right after the rename happens.  This
repo is not very active, so it's probably not worth doing a two step
approach of first adding the main branch renaming and then removing
the `master` reference from the workflow files.