install tap directly instead of via --formula
When using `brew install --formula ...`, homebrew complains that it
requires formulae to be in a tap, and rejects the installation. We do
however install the tap already in the previous step using `brew tap
pulumi/tap .`, so we don't need to use the formula directly.

`brew install pulumi/tap/pulumi` should still give us the same result
as installing the formula directly, but actually works in CI.